@@ -105,8 +105,8 @@ class DonchianChannelsSeries extends Series {
       return ChannelFillPainter(
         upperChannelSeries,
         lowerChannelSeries,
-        firstUpperChannelFillColor: config.fillColor.withOpacity(0.2),
-        secondUpperChannelFillColor: config.fillColor.withOpacity(0.2),
+        firstUpperChannelFillColor: config.fillColor.withValues(alpha: 0.2),
+        secondUpperChannelFillColor: config.fillColor.withValues(alpha: 0.2),
       );
     }
 
